Tidy and put the finishing touches on a model output

Wrapper around broom::tidy() with some additional finishing touches!

polish(
  x,
  .labels = NULL,
  .conf_int = TRUE,
  .flextable = TRUE,
  .header1 = NULL,
  .header2 = NULL,
  ...
)

Arguments

x
A model object to be converted into a tibble::tibble()

.labels
A named list of variable labels. If NULL, current labels will be used

.conf_int
Logical to return a 95 percent confidence interval

.flextable
Logical to return a formatted flextable object if TRUE or a list of a dataframe and row numbers to use in flextable::padding()

.header1
See ??mRclwhip::format_flextable for details

.header2
See ??mRclwhip::format_flextable for details

...
additional arguments passed to broom::tidy()

Examples

df1 <- tibble::tibble(
  id = letters,
  age = sample(seq(18, 39, 1), 26, replace = T),
  mile_time = sample(seq(6, 7.5, 0.1), 26, replace = T),
  gender = sample(c("Male", "Female"), 26, replace = T),
  country = sample(c("USA", "Canada", "Africa", "England"), 26, replace = T)
)

lm_res <- lm(mile_time ~ age + gender + country, data = df)

#> Error in as.data.frame.default(data, optional = TRUE): cannot coerce class ‘"function"’ to a data.frame

lm_res %>% polish(.labels = c(age = "Age", gender = "Gender", country = "Country of origin"), .header1 = list(values = c("", "Estimate (95% CI)")))

#> Error in eval(lhs, parent, parent): object 'lm_res' not found